refactor: move wizard/constants.ts to data/designConstants.ts (#42)
